requirements.txt
openai==0.28.1
python-dotenv==1.0.0
flask==3.0.2
flask-socketio==5.3.6
mysql-connector-python==8.0.33
werkzeug==3.0.1 